+++ crates/api-desc/CHANGELOG.md
@@ -24,4 +24,4 @@
 
 ## 0.1.0
 
-<!-- Update PR number to skip CHANGELOG.md test: #0 -->
+<!-- Update PR number to skip CHANGELOG.md test: #139 -->